+++ bxfinance-scenario1/src/components/Integration/PingData.js
